+++ components/shared/custome/c-drawer.tsx
@@ -5,31 +5,28 @@ import {
   DrawerTitle,
 } from "@/components/ui/drawer";
 import { cn } from "@/lib/utils";
+import CloseModal from "../close-dialog";
 
 const CDrawer = ({
   open,
   setIsOpen,
   children,
   headerTitle,
-  height,
   paddingContent = true,
+  showClose = true,
 }: CDrawer) => {
   return (
     <Drawer
       open={open}
       onOpenChange={(isOpen) => (!isOpen ? setIsOpen() : null)}
+      dismissible={false}
     >
-      <DrawerContent
-        className={cn(
-          "bg-neutral overflow-hidden !rounded-t-md",
-          height || "h-[300px]",
-        )}
-      >
-        {" "}
-        <DrawerHeader className="!p-0">
-          <DrawerTitle className="flex pb-5 px-4">{headerTitle}</DrawerTitle>
+      <DrawerContent>
+        <DrawerHeader className={cn(showClose && "text-left py-0")}>
+          <DrawerTitle className="text-foreground!">{headerTitle}</DrawerTitle>
+          {showClose && <CloseModal />}
         </DrawerHeader>
-        <div className={cn("p-4", !paddingContent && "!p-0")}>{children}</div>
+        <div className={cn("p-4", !paddingContent && "p-0")}>{children}</div>
       </DrawerContent>
     </Drawer>
   );